requirements.txt
Flask==2.0.1
pillow>=9.0.0
pdfkit>=1.0.0
werkzeug==2.0.1